Transition the primary when all its partitions have been stored/recovered (#8)
If all partitions from a given sample are stored, the primary sample is
transitioned to "stored" status automatically too. Partitions in status
"rejected", "cancelled" and "invalid" are not considered.

If all partitions from a given sample have been recovered, the primary
sample is automatically transitioned to its previous status (before it
was transitioned to "stored").